Refactored tests.
